% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDump.R
\name{dataDump}
\alias{dataDump}
\title{Dump all of curatedMetagenomicData as two .csv files}
\usage{
dataDump(dataType = "relative_abundance", counts = FALSE)
}
\arguments{
\item{dataType}{(default: "relative_abundance")
Data type, passed on to \code{\link[curatedMetagenomicData]{returnSamples}}.}

\item{counts}{(Default: FALSE)
Whether to convert to count-like data by multiplying through by read depth.
Passed on to \code{\link[curatedMetagenomicData]{returnSamples}}.}
}
\value{
The SummarizedExperiment or TreeSummarizedExperiment containing all of cMD for the data type requested.
Calling this function has the side-effect of also writing two csv files, one for the assay data of this
object and one for the colData.
}
\description{
Dump all of curatedMetagenomicData as two .csv files
}
\details{
This function also removes control samples from the NielsenHB_2014 study,
which are duplicated in the LeChatelierE_2013 study. The cMD version number is put in the filename.
}
\examples{
\dontrun{
   tse <- dataDump()
   tse
   dir(pattern = "\\\\.csv")
}
}
