% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sort.R
\name{filter_sort}
\alias{filter_sort}
\title{Filter and Sort Features in Meta-analysis final data-frame}
\usage{
filter_sort(df_to_sort, na_filter)
}
\arguments{
\item{df_to_sort}{The dataframe containing the meta-analysis results (final_df
in this vignette).}

\item{na_filter}{Maximum number of cohorts for which the SDM difference couldn't
be computed to allow for a feature to be included.}
}
\value{
The dataframe with the meta-analysis results, filtered for prevalent
features and ordered by adj.p-value and effect size.
}
\description{
This function allows removal from the dataframe with the meta-analysis
results those features for which the Standardized Mean Difference couldn't
be computed in at least a user-specified number of cohorts (na_filter).
}
\details{
Additionally, this function orders in increasing order the features according
to the corrected p-values and in decreasing order for the Overall effect size
}
