% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSEforCondition.R
\name{makeSEforCondition}
\alias{makeSEforCondition}
\title{Title Make a dataset for a condition of interest.}
\usage{
makeSEforCondition(
  condition,
  removestudies = NULL,
  dataType = "relative_abundance",
  counts = FALSE
)
}
\arguments{
\item{condition}{Condition of study for which to build a case-control dataset.
See "study_condition" column of the \code{sampleMetadata} object.}

\item{removestudies}{Any studies not to be included (default: NULL)}

\item{dataType}{Type of metagenomic data to return, see \code{?curatedMetagenomicData}}

\item{counts}{Convert to something resembling counts, by multiplying through
by read depth?}
}
\value{
a (Tree)SummarizedExperiment containing merged data from 1+ studies
}
\description{
Title Make a dataset for a condition of interest.
}
\details{
This function finds datasets that contain the condition of interest, returns
those datasets, and filters them to contain only samples of the condition or
controls. These datasets are then merged into a single
(Tree)SummarizedExperiment. Controls from other datasets are not included.
}
\examples{
makeSEforCondition("STH")
}
