% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaanalysis_quantitative.R
\name{runMetaanalysis_quantitative}
\alias{runMetaanalysis_quantitative}
\title{Single Species Meta-analysis (Quantitative Variables)}
\usage{
runMetaanalysis_quantitative(R_vector, n)
}
\arguments{
\item{R_vector}{A vector with correlations for a
single feature (Species) across datasets.}

\item{n}{A vector with the number of samples in each cohort}
}
\value{
A named vector with the results of the meta-analysis for the feature
implicitly described by d_vector and SE_d_vector.
RE_Correlation","SE_RE","CI_RE","Zscore","p-value","tau2","I^2"
}
\description{
Wrapping function to perform a meta-analysis using function \emph{metagen}
from package \emph{meta} for a single Species. Between-study variance is
computed with Paule-Mandel estimator (method.tau = PM). The summary measure
is the Standardize Mean Difference (sm = ZCOR)
}
