% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleDatasetAnalyze.R
\name{single_dataset_analyze}
\alias{single_dataset_analyze}
\title{Meta-analysis on a single dataset}
\usage{
single_dataset_analyze(exprs_df, meta_df, dataset_name)
}
\arguments{
\item{exprs_df}{relative abundance matrix with samples in column}

\item{meta_df}{metadata containing...}

\item{dataset_name}{}
}
\value{

}
\description{
Meta-analysis on a single dataset
}
